graph reordering + floats

+++ src/common/csr_utils.h
@@ -9,7 +9,7 @@ typedef struct {
     uint64_t n_rows;
     uint64_t n_columns;
     uint64_t nnz;
-    double *values;
+    float *values;
     uint64_t *columns;
     uint64_t *row_ptrs;
 } csr_matrix_t;
