+++ src/pagerank_seq.c
@@ -6,18 +6,18 @@
 #include <math.h>
 #include "common/csr_utils.h"
 
-bool pagerank(const csr_matrix_t *const mat, double **rank);
-bool pagerank_original(const csr_matrix_t *const mat, double **rank);
-void cleanup(csr_matrix_t *mat, double *rank);
+bool pagerank(const csr_matrix_t *const mat, float **rank);
+bool pagerank_original(const csr_matrix_t *const mat, float **rank);
+void cleanup(csr_matrix_t *mat, float *rank);
 
-void matvec_mul(const csr_matrix_t *const mat, const double *const vec, double *const out);
-void vec_add_scalar(const double *const vec1, double scalar, double *const out, uint64_t size);
-void vec_diff(const double *const vec1, const double *const vec2, double *const out, uint64_t size);
-void linear_comb(const double *const vec1, const double *const vec2, const double a1, const double a2, double *const out, uint64_t size);
-double l1_norm(const double *const vec, uint64_t size);
+void matvec_mul(const csr_matrix_t *const mat, const float *const vec, float *const out);
+void vec_add_scalar(const float *const vec1, float scalar, float *const out, uint64_t size);
+void vec_diff(const float *const vec1, const float *const vec2, float *const out, uint64_t size);
+void linear_comb(const float *const vec1, const float *const vec2, const float a1, const float a2, float *const out, uint64_t size);
+float l1_norm(const float *const vec, uint64_t size);
 
 void find_dangling(const csr_matrix_t *const mat, uint64_t *const out, uint64_t *out_size);
-double rank_loss(const uint64_t *const dangling_indices, const uint64_t dangling_size, const double *const rank, const uint64_t size);
+float rank_loss(const uint64_t *const dangling_indices, const uint64_t dangling_size, const float *const rank, const uint64_t size);
 
 
 int main(int argc, char **argv) {
@@ -34,7 +34,7 @@ int main(int argc, char **argv) {
     }
 
     // Run PageRank
-    double *rank = malloc(sizeof(double) * web.n_rows);
+    float *rank = malloc(sizeof(float) * web.n_rows);
     if (!pagerank(&web, &rank)) {
         cleanup(&web, rank);
         return -1;
@@ -50,39 +50,39 @@ int main(int argc, char **argv) {
 /**
  * PageRank implementation
  */
-bool pagerank(const csr_matrix_t *const mat, double **rank) {
+bool pagerank(const csr_matrix_t *const mat, float **rank) {
     if (mat->n_rows != mat->n_columns) {
         fprintf(stderr, "ERROR: n_rows (%" PRIu64 ") must be equal to n_columns (%" PRIu64 ")", mat->n_rows, mat->n_columns);
         return false;
     }
 
-    const double EPSILON = 1e-7;
-    const double GLOBAL_RANK = 1.0;
+    const float EPSILON = 1e-7;
+    const float GLOBAL_RANK = 1.0;
     const uint32_t MAX_ITERATIONS = 1000;
-    const double RANDOM_JUMP_PROB = 0.15;
+    const float RANDOM_JUMP_PROB = 0.15;
     
     uint64_t size = mat->n_rows;
 
     // Initial distribution: uniform
-    double *e = malloc(sizeof(double) * size);
+    float *e = malloc(sizeof(float) * size);
     for (uint64_t i = 0; i < size; i++) {
         e[i] = GLOBAL_RANK / size;
     }
 
     // Resource allocation
-    double *vec0 = malloc(sizeof(double) * size);
-    double *vec1 = malloc(sizeof(double) * size);
-    double *diff = malloc(sizeof(double) * size);
+    float *vec0 = malloc(sizeof(float) * size);
+    float *vec1 = malloc(sizeof(float) * size);
+    float *diff = malloc(sizeof(float) * size);
     uint64_t *dangling = malloc(sizeof(uint64_t) * size);
 
     // Initialization
-    double *last_rank = vec0;
-    double *new_rank = vec1;
-    memcpy(vec0, e, sizeof(double) * size);
+    float *last_rank = vec0;
+    float *new_rank = vec1;
+    memcpy(vec0, e, sizeof(float) * size);
     uint64_t dangling_size;
     find_dangling(mat, dangling, &dangling_size);
 
-    double delta;
+    float delta;
 
     uint32_t iteration = 0;
     do {
@@ -90,18 +90,18 @@ bool pagerank(const csr_matrix_t *const mat, double **rank) {
         matvec_mul(mat, last_rank, new_rank);
 
         // Compute lost rank (dangling pages)
-        double teleport = rank_loss(dangling, dangling_size, new_rank, size);
+        float teleport = rank_loss(dangling, dangling_size, new_rank, size);
         
         // Random surfer
         linear_comb(new_rank, e, 1.0 - RANDOM_JUMP_PROB, RANDOM_JUMP_PROB, new_rank, size);
-        vec_add_scalar(new_rank, teleport / (double)size, new_rank, size);
+        vec_add_scalar(new_rank, teleport / (float)size, new_rank, size);
 
         // Compute delta
         vec_diff(new_rank, last_rank, diff, size);
         delta = l1_norm(diff, size);
 
         // Update data of last iteration
-        double *aux = new_rank;
+        float *aux = new_rank;
         new_rank = last_rank;
         last_rank = aux;
 
@@ -111,7 +111,7 @@ bool pagerank(const csr_matrix_t *const mat, double **rank) {
     } while(delta > EPSILON && iteration < MAX_ITERATIONS);
 
     // Store result
-    memcpy(*rank, last_rank, sizeof(double) * size);
+    memcpy(*rank, last_rank, sizeof(float) * size);
 
     // Cleanup
     free(e);
@@ -126,7 +126,7 @@ bool pagerank(const csr_matrix_t *const mat, double **rank) {
 /**
  * Frees the resources used by the main function
  */
-void cleanup(csr_matrix_t *mat, double *rank) {
+void cleanup(csr_matrix_t *mat, float *rank) {
     free(rank);
     free(mat->values);
     free(mat->columns);
@@ -136,9 +136,9 @@ void cleanup(csr_matrix_t *mat, double *rank) {
 
 // Helper functions
 
-void matvec_mul(const csr_matrix_t *const mat, const double *const vec, double *const out) {
+void matvec_mul(const csr_matrix_t *const mat, const float *const vec, float *const out) {
     for (uint64_t j = 0; j < mat->n_rows; j++) {
-        double sum = 0.0;
+        float sum = 0.0;
 
         for (uint64_t index = mat->row_ptrs[j]; index < mat->row_ptrs[j + 1]; index++) {
             sum += mat->values[index] * vec[mat->columns[index]];
@@ -148,26 +148,26 @@ void matvec_mul(const csr_matrix_t *const mat, const double *const vec, double *
     }
 }
 
-void vec_add_scalar(const double *const vec1, double scalar, double *const out, uint64_t size) {
+void vec_add_scalar(const float *const vec1, float scalar, float *const out, uint64_t size) {
     for (uint64_t i = 0; i < size; i++) {
         out[i] = vec1[i] + scalar;
     }
 }
 
-void vec_diff(const double *const vec1, const double *const vec2, double *const out, uint64_t size) {
+void vec_diff(const float *const vec1, const float *const vec2, float *const out, uint64_t size) {
     for (uint64_t i = 0; i < size; i++) {
         out[i] = vec1[i] - vec2[i];
     }
 }
 
-void linear_comb(const double *const vec1, const double *const vec2, const double a1, const double a2, double *const out, uint64_t size) {
+void linear_comb(const float *const vec1, const float *const vec2, const float a1, const float a2, float *const out, uint64_t size) {
     for (uint64_t i = 0; i < size; i++) {
         out[i] = vec1[i] * a1 + vec2[i] * a2;
     }
 }
 
-double l1_norm(const double *const vec, uint64_t size) {
-    double sum = 0.0;
+float l1_norm(const float *const vec, uint64_t size) {
+    float sum = 0.0;
     for (uint64_t i = 0; i < size; i++) {
         sum += fabs(vec[i]);
     }
@@ -190,8 +190,8 @@ void find_dangling(const csr_matrix_t *const mat, uint64_t *const out, uint64_t 
     }
 }
 
-double rank_loss(const uint64_t *const dangling_indices, const uint64_t dangling_size, const double *const rank, const uint64_t size) {
-    double sum = 0.0;
+float rank_loss(const uint64_t *const dangling_indices, const uint64_t dangling_size, const float *const rank, const uint64_t size) {
+    float sum = 0.0;
     for (uint64_t i = 0; i < dangling_size; i++) {
         sum += rank[dangling_indices[i]];
     }
@@ -203,35 +203,35 @@ double rank_loss(const uint64_t *const dangling_indices, const uint64_t dangling
  * Original PageRank implementation
  * Discarded since it converges slowly due to the model used for the random surfer
  */
-bool pagerank_original(const csr_matrix_t *const mat, double **rank) {
+bool pagerank_original(const csr_matrix_t *const mat, float **rank) {
     if (mat->n_rows != mat->n_columns) {
         fprintf(stderr, "ERROR: n_rows (%" PRIu64 ") must be equal to n_columns (%" PRIu64 ")", mat->n_rows, mat->n_columns);
         return false;
     }
 
-    const double EPSILON = 1e-7;
-    const double GLOBAL_RANK = 1.0;
+    const float EPSILON = 1e-7;
+    const float GLOBAL_RANK = 1.0;
     const uint32_t MAX_ITERATIONS = 1000;
     
     uint64_t size = mat->n_rows;
 
     // Initial distribution: uniform
-    double *e = malloc(sizeof(double) * size);
+    float *e = malloc(sizeof(float) * size);
     for (uint64_t i = 0; i < size; i++) {
         e[i] = GLOBAL_RANK / size;
     }
 
     // Resource allocation
-    double *vec0 = malloc(sizeof(double) * size);
-    double *vec1 = malloc(sizeof(double) * size);
-    double *diff = malloc(sizeof(double) * size);
+    float *vec0 = malloc(sizeof(float) * size);
+    float *vec1 = malloc(sizeof(float) * size);
+    float *diff = malloc(sizeof(float) * size);
 
     // Initialization
-    double *last_rank = vec0;
-    double *new_rank = vec1;
-    memcpy(vec0, e, sizeof(double) * size);
+    float *last_rank = vec0;
+    float *new_rank = vec1;
+    memcpy(vec0, e, sizeof(float) * size);
 
-    double delta;
+    float delta;
 
     uint32_t iteration = 0;
     do {
@@ -239,9 +239,9 @@ bool pagerank_original(const csr_matrix_t *const mat, double **rank) {
         matvec_mul(mat, last_rank, new_rank);
         
         // Random surfer
-        double last_norm = l1_norm(last_rank, size);
-        double new_norm = l1_norm(new_rank, size);
-        double d = last_norm - new_norm;
+        float last_norm = l1_norm(last_rank, size);
+        float new_norm = l1_norm(new_rank, size);
+        float d = last_norm - new_norm;
         printf("d=%lf\n", d);
         linear_comb(new_rank, e, 1.0, d, new_rank, size);
 
@@ -250,7 +250,7 @@ bool pagerank_original(const csr_matrix_t *const mat, double **rank) {
         delta = l1_norm(diff, size);
 
         // Update data of last iteration
-        double *aux = new_rank;
+        float *aux = new_rank;
         new_rank = last_rank;
         last_rank = aux;
 
@@ -261,7 +261,7 @@ bool pagerank_original(const csr_matrix_t *const mat, double **rank) {
     } while(delta > EPSILON && iteration < MAX_ITERATIONS);
 
     // Store result
-    memcpy(*rank, last_rank, sizeof(double) * size);
+    memcpy(*rank, last_rank, sizeof(float) * size);
 
     // Cleanup
     free(e);
